Fixed types for MinGW

+++ src/strsearch_enc_1.cpp
@@ -38,7 +38,7 @@ void get_input_from_term(vecChar& a) {
   std::getline(std::cin, cstr);
   cout <<"Pattern is `"<<cstr<<"'"<<endl;    
   for(auto c: cstr) {
-	a.push_back(c);
+	  a.push_back(c);
   }
   cout <<"Pattern is "<<a.size()<<" characters"<<endl;  
   return;
@@ -51,12 +51,12 @@ void get_input_from_file(vecChar& a, string fname) {
   ifstream in_file;
   in_file.open(fname);
   if (!in_file) {
-	cerr << "Can't open file for input: "<<fname;
-	exit(-1); //error exit
+    cerr << "Can't open file for input: "<<fname;
+    exit(-1); //error exit
   }
   
   while (in_file >> c) {
-	a.push_back(c);
+	  a.push_back(c);
   }
   cout <<"Read "<<a.size()<<" characters"<<endl;
   in_file.close();
@@ -65,64 +65,64 @@ void get_input_from_file(vecChar& a, string fname) {
      
 // plaintext string search of pat within txt, with modulus of ps          
 vecInt search(vecChar &pat, vecChar &txt, int ps) {
-  long p(ps);
+  int64_t p(ps);
   DEBUG_FLAG(false);
-  int M = pat.size();
+  size_t M = pat.size();
   DEBUGEXP(M);
-  int N = txt.size();
+  size_t N = txt.size();
   DEBUGEXP(N);
-  int i, j;
-  long ph = 0;  // hash value for pattern
-  long th = 0; // hash value for txt
-  long h = 1;
+  size_t i, j;
+  int64_t ph = 0;  // hash value for pattern
+  int64_t th = 0; // hash value for txt
+  int64_t h = 1;
 
-  int nfound = 0;
+  size_t nfound = 0;
      
   // The value of h would be "pow(d, M-1)%p"
   for (i = 0; i < M-1; i++) {
-	h = (h*d)%p;
-	DEBUGEXP(h);
+    h = (h*d)%p;
+    DEBUGEXP(h);
   }
   DEBUG(" hfinal: "<<h);
 
   // Calculate the hash value of pattern and first window of text
   for (i = 0; i < M; i++) {
-	ph = (d * ph + pat[i]) % p;
-	th = (d * th + txt[i]) % p;
+    ph = (d * ph + pat[i]) % p;
+    th = (d * th + txt[i]) % p;
   }
   DEBUG(" initial ph: "<<ph);
   DEBUG(" initial th: "<<th);
   vecInt pres(0);
   // Slide the pattern over text one by one
   for (i = 0; i <= N - M; i++) {
-	
-	// Check the hash values of current window of text and pattern
-	// If the hash values match then only check for characters on by one
-	pres.push_back((ph-th)%p);
-	if ( ph == th )	{
-	  /* Check for characters one by one */
-	  for (j = 0; j < M; j++) {
-		if (txt[i + j] != pat[j])
-		  break;
-	  }
-	  if (j == M) { // if ph == t and pat[0...M-1] = txt[i, i+1, ...i+M-1]
+    
+    // Check the hash values of current window of text and pattern
+    // If the hash values match then only check for characters on by one
+    pres.push_back((ph-th)%p);
+    if ( ph == th )	{
+      /* Check for characters one by one */
+      for (j = 0; j < M; j++) {
+      if (txt[i + j] != pat[j])
+        break;
+      }
+      if (j == M) { // if ph == t and pat[0...M-1] = txt[i, i+1, ...i+M-1]
 
-		cout<<"Pattern found at index "<< i << endl;
-		nfound++;
-	  }
-	}
-     
-	// Calculate hash value for next window of text: Remove leading digit,
-	// add trailing digit
-	if ( i < N - M ) {
-	  th = (d * (th - txt[i] * h) + txt[i + M]) % p;
+      cout<<"Pattern found at index "<< i << endl;
+      nfound++;
+      }
+    }
+      
+    // Calculate hash value for next window of text: Remove leading digit,
+    // add trailing digit
+    if ( i < N - M ) {
+      th = (d * (th - txt[i] * h) + txt[i + M]) % p;
 
-	  // We might get negative value of t, converting it to positive
-	  if (th < 0) {
-		th = (th + p);
-	  }
+      // We might get negative value of t, converting it to positive
+      if (th < 0) {
+      th = (th + p);
+      }
 
-	}
+    }
   } //end for
 
   cout<<"total occurances " <<nfound<<endl;
@@ -143,12 +143,12 @@ CT encrypt_repeated_integer(CryptoContext<DCRTPoly> &cc, LPPublicKey<DCRTPoly> &
 // helper function to multiply by constant 256 using binary tree addition
 CT encMultD(CryptoContext<DCRTPoly> &cc, CT in){
   if (d !=256){
-	cout <<"error d not 256"<<endl;
-	exit(-1);
+    cout <<"error d not 256"<<endl;
+    exit(-1);
   }
   auto tmp(in);
   for (auto i = 0; i< 8; i++ ){
-	tmp = cc->EvalAdd(tmp, tmp);
+	  tmp = cc->EvalAdd(tmp, tmp);
   }
   
   return(tmp);
@@ -157,13 +157,13 @@ CT encMultD(CryptoContext<DCRTPoly> &cc, CT in){
 //Single value encrypted search     
 vecCT encrypted_search(CryptoContext<DCRTPoly> &cc,  LPPublicKey<DCRTPoly> &pk, vecCT &epat, vecCT &etxt, int ps) {
 
-  long p(ps);
+  int64_t p(ps);
   DEBUG_FLAG(false);
-  int M = epat.size();
+  size_t M = epat.size();
   DEBUGEXP(M);
-  int N = etxt.size();
+  size_t N = etxt.size();
   DEBUGEXP(N);
-  int i;
+  size_t i;
 
   PT dummy;
   
@@ -174,20 +174,20 @@ vecCT encrypted_search(CryptoContext<DCRTPoly> &cc,  LPPublicKey<DCRTPoly> &pk, 
 
   DEBUG("encrypting hct");     
   // The value of h would be "pow(d, M-1)%p"
-  long h = 1;
+  int64_t h = 1;
   for (i = 0; i < M-1; i++) {
-	h = (h*d)%p;
+	  h = (h*d)%p;
   }
   CT hct = encrypt_repeated_integer(cc, pk, h, nrep);  // encrypted h
 
   DEBUG("encrypting first hashes" );     
   // Calculate the hash value of pattern and first window of text
   for (i = 0; i < M; i++) {
-	auto tmp = encMultD(cc, phct);	
-	phct = cc->EvalAdd(tmp, epat[i]);
+    auto tmp = encMultD(cc, phct);	
+    phct = cc->EvalAdd(tmp, epat[i]);
 
-	tmp = encMultD(cc, thct);
-	thct = cc->EvalAdd(tmp, etxt[i]);
+    tmp = encMultD(cc, thct);
+    thct = cc->EvalAdd(tmp, etxt[i]);
   }
 
   vecCT eres(0);
@@ -235,10 +235,10 @@ int main()
   get_input_from_file(bigtxt, infilename);
 
   //cout<<"Enter text size:";
-  unsigned int textSize(0);
+  uint32_t textSize(0);
   //cin>> textSize;
   textSize = 32;
-  unsigned int offset(16);
+  uint32_t offset(16);
   cout << "Limiting search to "<<textSize<< " characters "
 	  <<"starting at offset "<<offset<<endl;
 
@@ -294,36 +294,36 @@ int main()
   //encrypt the pattern
   vecInt vin(0);
   vecCT epat(0);
-  unsigned int j(0);
+  uint32_t j(0);
   for (auto ch: pat) {
-	cout<<j<< '\r'<<flush;
-	j++;
-	vin.push_back(ch);
-	PT pt= cc->MakePackedPlaintext(vin);
-	vin.clear();
-	CT ct = cc->Encrypt(keyPair.publicKey, pt);
-	epat.push_back(ct);
+    cout<<j<< '\r'<<flush;
+    j++;
+    vin.push_back(ch);
+    PT pt= cc->MakePackedPlaintext(vin);
+    vin.clear();
+    CT ct = cc->Encrypt(keyPair.publicKey, pt);
+    epat.push_back(ct);
   }	
 
   //encrypt the text
   auto ringsize = cc->GetRingDimension();
   cout << "ringsize = "<<ringsize << endl;
   cout << "txt size = "<<txt.size() << endl;
-  unsigned int nbatch = int(ceil(float(txt.size())/float(ringsize)));
+  uint32_t nbatch = int(ceil(float(txt.size())/float(ringsize)));
   cout << "can store "<<nbatch <<" batches in the ct"<<endl;
 	
   cout<<"Step 3.2 - Encrypt text"<<endl;  
   vecCT etxt(0);
   auto pt_len(0);
   for (usint i = 0; i < txt.size(); i++) {
-	cout<<i<< '\r'<<flush;
+    cout<<i<< '\r'<<flush;
 
-	vin.push_back(txt[i]);
-	Plaintext pt= cc->MakePackedPlaintext(vin);
-	pt_len = pt->GetLength();
-	vin.clear();
-	CT ct = cc->Encrypt(keyPair.publicKey, pt);
-	etxt.push_back(ct);
+    vin.push_back(txt[i]);
+    Plaintext pt= cc->MakePackedPlaintext(vin);
+    pt_len = pt->GetLength();
+    vin.clear();
+    CT ct = cc->Encrypt(keyPair.publicKey, pt);
+    etxt.push_back(ct);
   }	
   cout<<"Step 4 - Encrypted string search"<<endl;  
 
@@ -336,21 +336,21 @@ int main()
   //decrypt the result and compute location of potential matches
   vecPT vecResult(0);
   for (auto e_itr:eresult){
-	PT ptresult;  
-	cc->Decrypt(keyPair.secretKey, e_itr, &ptresult);
-	ptresult->SetLength(pt_len);
-	vecResult.push_back(ptresult);
+    PT ptresult;  
+    cc->Decrypt(keyPair.secretKey, e_itr, &ptresult);
+    ptresult->SetLength(pt_len);
+    vecResult.push_back(ptresult);
   }
 
   int i(0);
   int nfound(0);
   for (auto val: vecResult) {
-	auto unpackedVal = val->GetPackedValue();
-	if (unpackedVal[0] == 0) {
-	  cout<<"Pattern found at index "<< i << endl;
-	  nfound++;
-	}
-	i++;
+    auto unpackedVal = val->GetPackedValue();
+    if (unpackedVal[0] == 0) {
+      cout<<"Pattern found at index "<< i << endl;
+      nfound++;
+    }
+    i++;
   }
   cout<<"total occurances "<<nfound<<endl;
